moved tokenzizer and vocab_size to the base model
