+++ nemo/collections/nlp/models/language_modeling/megatron_base_model.py
@@ -19,6 +19,8 @@ from pytorch_lightning.trainer.trainer import Trainer
 
 from nemo.collections.nlp.models.nlp_model import NLPModel
 from nemo.collections.nlp.modules.common.megatron.megatron_init import initialize_model_parallel_for_nemo
+from nemo.collections.nlp.modules.common.tokenizer_utils import get_nmt_tokenizer
+from nemo.utils import logging
 
 try:
     import apex
@@ -74,6 +76,12 @@ class MegatronBaseModel(NLPModel):
             apex_transformer_log_level=self.cfg.get('apex_transformer_log_level', 30),
         )
 
+        # build tokenizer (defaults to nemo supported tokenizers)
+        self._build_tokenizer()
+
+        # manipulate vocabulary (e.g., pad vocabulary for better efficiency)
+        self._build_vocab()
+
     def _enable_nvidia_optimizations(self):
         "These optimizations are present in NVIDIA NGC PyTorch Containers"
 
@@ -102,3 +110,49 @@ class MegatronBaseModel(NLPModel):
         else:
             # Not a Nvidia container. Dependency check is on users
             pass
+
+    def _build_tokenizer(self):
+        """
+        Default tokenizer is based on available nemo tokenizers.
+        Override this method to use an external tokenizer.
+        All tokenizers are expected to provide compatible interface.
+        Override default Encoder-decoder tokenizer to use legacy=True for sentencepiece.
+        """
+        self.tokenizer = get_nmt_tokenizer(
+            library=self._cfg.tokenizer.library,
+            model_name=self._cfg.tokenizer.type,
+            tokenizer_model=self.register_artifact("tokenizer.model", self._cfg.tokenizer.model),
+            vocab_file=self.register_artifact("tokenizer.vocab_file", self._cfg.tokenizer.vocab_file),
+            merges_file=self.register_artifact("tokenizer.merge_file", self._cfg.tokenizer.merge_file),
+            delimiter=self.cfg.tokenizer.get('delimiter', None),
+            legacy=True if self._cfg.tokenizer.library == 'sentencepiece' else False,
+        )
+        # add pad special token
+        if not hasattr(self.tokenizer, "pad_id"):
+            self.tokenizer.add_special_tokens({'pad_token': '<pad>'})
+        elif hasattr(self.tokenizer, "pad_id") and (self.tokenizer.pad_id is None or self.tokenizer.pad_id < 0):
+            self.tokenizer.add_special_tokens({'pad_token': '<pad>'})
+
+    def _build_vocab(self):
+        """
+        Manipulate vocabulary (e.g., pad vocabulary for increased performance)/
+        """
+        # TODO: add config to allow to disable it?
+        self.padded_vocab_size = self._vocab_size_with_padding(
+            orig_vocab_size=self.tokenizer.vocab_size,
+            make_vocab_size_divisible_by=self._cfg.get('make_vocab_size_divisible_by', 128),
+            tensor_model_parallel_size=self._cfg.get('tensor_model_parallel_size', 1),
+        )
+
+    def _vocab_size_with_padding(self, orig_vocab_size, make_vocab_size_divisible_by, tensor_model_parallel_size):
+        """Pad vocab size so it is divisible by model parallel size and
+        still having GPU friendly size."""
+
+        after = orig_vocab_size
+        multiple = make_vocab_size_divisible_by * tensor_model_parallel_size
+        while (after % multiple) != 0:
+            after += 1
+        logging.info(
+            f'Padded vocab_size: {after}, original vocab_size: {orig_vocab_size}, dummy tokens: {after - orig_vocab_size}.'
+        )
+        return after
